Third version of the Django project
When the submit button is pressed, a check is performed to ensure the validity of the URL, which must correspond to a property page on Booking.com. Then, we take this URL and use Apify's Booking.com Reviews Scraper API to retrieve the reviews of a property listed on Booking.com and print them to the console.