Fourth version of the Django project
Using the Booking.com Reviews Scraper API from Apify, we retrieve 150 reviews from an accommodation provided by the user. These reviews are then used and analyzed by a machine learning model, in our case GPT-4o-mini, to generate a summary of the accommodation's reviews. This summary, which is approximately 150 words, is then printed to the webpage.