Fifth version
Beyond the summary of the reviews, the user also has the ability to receive answers to specific questions, which are based on the entirety of the reviews that have been retrieved. The answers, of course, are displayed on the page. We use the GPT-4o-mini for the data analysis and conclusion extraction.