Seventh version
The function get_rev2 was added, which implements the retrieval of reviews from an accommodation on Booking.com using the Booking.com Reviews Scraper API by Caprolok. Additionally, the get_rev3 function was added, which performs the same task via a free scraper. This scraper is 5 to 6 times faster than the two APIs we have used within the scope of this project.